Moving APInt and APFloat helper functions to MLIUtils.h
