+++ lib/CMakeLists.txt
@@ -24,6 +24,7 @@ add_library(MLIRInterpreterLib SHARED
   Interpreter/InterpreterOpInterface.cpp
   Interpreter/Dialects/Func/FuncInterpreter.cpp
   Interpreter/Dialects/LLVM/LLVMInterpreter.cpp
+  ${CMAKE_SOURCE_DIR}/src/MLIUtils.cpp
 )
 
 # Include directories for the library
